cache_utils: promote batch_idx to int64 in gather-output offset (latent 1M-context overflow)
The output-row offset at the dequant-gather kernel computes
batch_idx * out_stride0 in int32 (tl.program_id is i32; out_stride0 binds as
i32). With the gather buffer sized for very long sequences
(out_stride0 = max_num_tokens * 512), the product wraps past 2^31 at
>= ~819K tokens with batch_idx >= 4, turning the store offset negative
(~-3.2 GB) — an out-of-bounds write. The sibling physical_block_idx multiply
a few lines above is already promoted with a comment naming this exact
overflow class; this site was missed.

Latent at the 58-142K contexts of the issue-#2 crashes (max product ~3.6e8),
but reachable within this recipe's advertised 1M window. Found by an
int32-offset audit of all deepseek_v4 kernel wrappers; verified by threshold
analysis (2^31 / (max_num_tokens*512) = 4.0 batches at exactly 2^20 tokens).

+++ recipe/overlay/vllm/models/deepseek_v4/common/ops/cache_utils.py
@@ -259,8 +259,16 @@ def _dequantize_and_gather_k_kernel(
         token_fp8_ptr = token_data_ptr
         token_bf16_ptr = token_data_ptr + fp8_dim
 
-        # Output pointer for this token (flattened)
-        output_row_ptr = out_ptr + batch_idx * out_stride0 + (offset + i) * out_stride1
+        # Output pointer for this token (flattened).
+        # int64: batch_idx * out_stride0 can exceed 2^31 when the gather buffer
+        # is sized for very long sequences (out_stride0 = max_num_tokens * 512;
+        # wraps at >= ~819K tokens with batch_idx >= 4) — same overflow class
+        # as the guarded physical_block_idx multiply above.
+        output_row_ptr = (
+            out_ptr
+            + batch_idx.to(tl.int64) * out_stride0
+            + (offset + i) * out_stride1
+        )
 
         # ========== Dequantize FP8 portion using UE8M0 ==========
         for qblock_idx in tl.static_range(n_quant_blocks):
